Proxy UI fix in Firefox && upgrade to Selenium 4.8.1.
